+++ CHANGES.md
@@ -21,8 +21,12 @@ To be released.
 
 ### Bug fixes
 
+-  Fixed a bug that `Swarm` reported `TaskCanceledException` as an unknown
+   exception while stopping. [[#275]]
+
 [#270]: https://github.com/planetarium/libplanet/pull/270
 [#273]: https://github.com/planetarium/libplanet/issues/273
+[#275]: https://github.com/planetarium/libplanet/pull/275
 [#276]: https://github.com/planetarium/libplanet/pull/276